untaint cwd when constructing full perl file path
Previously, we would load perl files using a possibly relative path,
which would work based on . being in @INC.  Now that . is being removed,
we need to use an absolute path (or ./ relative path).  When
absolutizing, cwd would be used, which can be tained.  Although loading
based on . in @INC violates the spirit of taint, it had previously
worked and a downstream module expected it.  Untaint the cwd that gets
used so we can load relative paths under taint mode.